update delete blog

+++ database/migrations/2020_06_24_171436_creat_blogs_table.php
@@ -14,7 +14,7 @@ class CreatBlogsTable extends Migration
     public function up()
     {
         Schema::create('blogs', function (Blueprint $table) {
-            $table->increments('id');
+            $table->unsignedBigInteger('id');
             $table->string('title');
             $table->longText('content');
             $table->string('author');
